+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.7
+
+Add `Future getPermissions(onGranted:onDenied:)`
+Add `Future<PermissionStatus> checkPermissions()`
+
 ## 0.0.6
 
 Allow Significant-change for longer location updates (iOS)